Removing volatile types from the implementation and documentation
We recently decided to retire volatile types but never got around to
doing it.  The rationale was that atomic types are what we really want
(and that, as a fallback, we should support bettern extern type
declarations that are preserved to codegen at which point a user could
create an extern type to get at C's volatile types if they wanted).

Now that we've got atomics, I finally got around to pulling out this
code (removing dead code is fun!)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19867 3a8e244f-b0f2-452b-bcba-4c88e055c3ca