Switched the release version of HPL to using DimensionalDist2D with BlockCyclicDim and ReplicatedDim; augmented DimensionalDist2D with DimensionalArr.dsiReindex(); modified printLocalesTasks() in HPCCProblemSize.chpl to print locale grid dimensions, optionally.
This conversion of HPL favored elegance over performance.
Cf. test/studies/hpcc/HPL/vass/hpl.chpl - is performance-oriented.


To improve the performance, could do the following (of which, most
non-compiler mods are done in test/studies/hpcc/HPL/vass/hpl.chpl):

* Allocate replA/replB (and their domains) just once at the beginning,
rather than for each call to schurComplement() - to reduce the overhead
of allocation/deallocation and privatization.

* Place Ab, replA, ReplB in the global scope and do not pass them around
as arguments. Or fix the compiler bug whereas in the generated code
the following code (or a variant thereof) results in a remote reference
for A(i) (to read A's pid) even when 'A' is privatized and A(i) is local:

    proc example(A) {
      on (a remote locale) do A(i);
    }

* Make the computation of Rest.dim(1)(row..#blkSize) and ...dim(2)...
in schurComplement() local, by computing them outside the 'local'
clause - or by fixing the compiler?

* Enable 'local' in schurComplement() after doing the above.

* Hoisting the above slicing of Rest from inner loopnest(s).

* Try various permutations of the 3 for loops in schurComplement()
to see which is the fastest.

* replicateD1()/replicateD2() always replicate the whole row/column.
Once our slicing becomes cheaper, we can consider
replicating only the needed data during the last several
iterations in LUFactorize. For a large problem size, though,
we expect this to offer only a small performance improvement.

  The version in test/studies/hpcc/HPL/vass creates the four dimension
specifiers {dim1, dim2} * {block-cyclic, replicated} just once then
uses them when creating various Dimensional-mapped arrays. The version
being committed creates those specifiers right in the 'dmapped' clauses.
The advantage of the former is ensuring that they [the specifiers] are
consistent among themselves and among their uses.
(I don't think performance is a consideration here.)


Other notes:

* Switched the domain map for 'Ab', 'MatVectSpace' et al.
from the multi-dimensional BlockCyclic to DimensionalDist2D
with BlockCyclicDim in each dimension. The motivations were
to try out DimensionalDist2D and ensuring consistency in
index-to-locale mapping with replA/replB?

* Inlined dgemmNativeInds(), primarily to avoid figuring out
how to make it work with the replicated arrays. Part of it
is that I felt it was simpler to have the smaller dimension
of the replicated arrays to be 1..blkSize, rather than
(some starting point)..#blkSize. Or maybe because this makes
for fewer array slice expressions, which are currently expensive.

* Only schurComplement() takes advantage of replication.
The overhead of replication currently makes it unprofitable
for panelSolve() and updateBlockRow().

* Hoisted into separate functions the creation of replA, replB and the
replication of Ab into replA/replB - so that the top-level code is tidier
(e.g. main(), LUFactorize()). But see performance discussion above.

* Explicitly computed the dimensions of the locale grid to compute on
(gridRows, gridCols). That's because each dimension specifier constructor
is currently set up so that it needs the extent of its locale grid dimension.
Also the DimensionalDist2D constructor needs the locale grid array,
so Locales is reshaped into targetLocales explicitly, too.

  Again, hoisted computations into helper functions to keep top-level
code tidier.

  When computing gridRows and gridCols, the grid is either rectangular
or with second dimension larger than the first dimension.
This is because others tend to use grids N * (2*N).

* Added a condition to invoke schurComplement() only when 'br' aka 'Rest'
is non-empty. This is (a) to avoid useless replication and (b) because
replication currently assumes that the row/column being replicated
are blkSize-high/wide. This assumption could be fixed, but it seems
easier just to provide the conditional.

* Fixed a couple of comments.


Implemented aliasing for DimensionalDist2D-mapped arrays.

  This is so that I could write 'const replA => replicateD2(Ab...)'
in HPL. That's because 'const replA = replicateD2(Ab...)' would be
less efficient (it would entail copying, not aliasing)
and incorrect to begin with (it would copy only one replicand
of the r.h.s., leaving the other replicands of the l.h.s.
unassigned). We could avoid aliasing but at the expense
of some loss of code elegance.

  Aliasing is implemented by proc DimensionalArr.dsiReindex().
It currently works only for aliasing, and will generate
a run-time error upon a general reindexing of
a DimensionalDist2D-mapped array.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19874 3a8e244f-b0f2-452b-bcba-4c88e055c3ca