Avoid overflow for int(8) case.
While looking this over while writing the atomics primer, I noticed
that I originally wrote one of the examples with max(uint(8)) in mind,
but didn't update it when I added the signed int types.  It continued
to succeed nightly testing due to convenient wrap around properties,
but was actually overflowing for int(8).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19877 3a8e244f-b0f2-452b-bcba-4c88e055c3ca