Add 'flag' as a reserved word (temporarily)
Due to the use of C macros when declaring the various atomic functions
in all (2) varieties of chpl-atomics.h, I had to put a typedef in
chpl-atomics.h for 'flag'.  This unfortunately will conflict with any
use of 'flag' in a Chapel program.

After an email discussion with Brad, we decided that given today's
freeze, the best course is to add it to the list of
reservedSymbolNames that get automatically mangled by the compiler
(one line change).

Ultimately, we'll remove flag from the reserved list and avoid using
the macros for declaring the atomic functions for booleans.  FYI,
using 'atomicbool' instead of 'atomicflag' would require avoiding the
macros as well as changes in the atomics module.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19879 3a8e244f-b0f2-452b-bcba-4c88e055c3ca