Enabled 64-bit (blockSize*numLocales) in BlockCyclicDim; improved the cannot-change-range-stride-for-BlockCyclicDim warning.
The first issue is this: Before, blockSize*numLocales for BlockCyclicDim
had to fit in 32 bits because it was stored in cycSizeT = uint(32).
This was kinda not cool.

Simpler fixes could have been:
- Just make cycSizeT 'int(64)'. I rejected that because that would add
  one branch to each call to mod() and divfloor(), which are used often.
- Make cycSizeT 'uint(64)'. That did not work out-of-the-box because
  it resulted in mod, divfloor, as well as arithmetic ops like /,
  to run on (int(64), uint(64)), which is currently unsupported.

Therefore I made cycSizeT 'int(64)', but provided versions of mod
and divfloor that eliminate the extra branch by *assuming* that
the second argument is positive.
  These are called modP2 and divfloorP2. I added a test, tmath.chpl,
that ensures their correctness (against mod/divfloor from Math).

I also renamed cycSizeT to bcdPosInt, for BlockCyclicDim-PositiveInteger.
Also replaced cycSizeTuser with plain 'int', for simplicity.


The second issue arises when a DimensionalDist2D-mapped array
changes its index set via an assignment to its domain.
If there is a BlockCyclicDim specifier, in some cases
the array contents in the intersection of the old and new index sets
will not be preserved. This violates a language requirement.

To address that, there has been a warning in BlockCyclicDim issued in
such a case. This commit issues the warning only if the domain being
reassigned has some arrays associated with it.
We could go further and  omit this warning if the aforementioned intersection
is empty.

Making it a fatal error, i.e. halt(), seemed too drastic,
so it is a warning for now.

There does not seem to be a convention for issuing run-time warnings,
so it simply prints to stderr.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19889 3a8e244f-b0f2-452b-bcba-4c88e055c3ca