Rewrite to fix the materials array
This rewrite changes the materials array from a dense array of
size Elems to a sparse subdomain of Elems to better represent
that some subdomain of Elems stores the given material.  This
worked prior to this change simply due to the fact that LULESH,
as defined, stores the material in every element.  But this
wasn't in the spirit of the benchmark which ought to store
materials only in some places.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19892 3a8e244f-b0f2-452b-bcba-4c88e055c3ca