Pushing ssca2/ and hpcc/ release examples directories into benchmarks/
[reviewed by Sung]

This commit pushes ssca2 and hpcc in the release examples directory
down one level into a new benchmarks/ directory.  READMEs and
Makefiles are updated and added to reflect the new structure.  The
main portion of this change is that tests outside of the release
that refer back to the release source files had to have their
paths updated.

Note that I set up the Makefiles and READMEs to refer to a lulesh
subdirectory which doesn't exist here yet.  That's my next commit.  I
also haven't done a full test on this yet to make sure I didn't miss
any other references back to the old locations, but am about to fire
that off now.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19899 3a8e244f-b0f2-452b-bcba-4c88e055c3ca