Remove hook for using native qthread sync vars.
This commit removes qthreads from the check in isSimpleSyncBaseType(),
because there appears to be a bug in the floating point
implementations of the default native qthreads sync implementation.
Given that we are past the code freeze date for our release, this
seemed the best course to take.

This was discovered when SSCA2 started getting the wrong answers.  The
small test case committed with this change also fails when using the
qthreads native syncs.  It appears that the decimal portion is lost in
updates to the value.

The test is self-testing, but could run into issues with comparison of
floating point numbers.  FWIW, I ran the test in a loop for 20 minutes
and didn't get an false negatives (for the default config var values).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19901 3a8e244f-b0f2-452b-bcba-4c88e055c3ca