Moving LULESH to test/release/examples/benchmarks/lulesh
Mostly, this is a move from:

  studies/lulesh/bradc   -> release/examples/benchmarks/lulesh
  studies/lulesh/lmeshes -> release/examples/benchmarks/lulesh/lmeshes

I also updated the READMEs to make them a bit more end-user-oriented
and added a Makefile.  I also had to update my lulesh-eof* tests to
deal with the fact that the lmeshes locations changed, but did this
by moving them into a new 'tests' subdirectory of my lulesh workspace
and gave them a local .lmesh file to work with (so that they wouldn't
have to track changes to the .lmesh file format to keep working).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19902 3a8e244f-b0f2-452b-bcba-4c88e055c3ca