Updating REGRESSIONS based on last night's runs
Things were noisy due to path changes for the release/examples
benchmarks, but the only real failure that I picked up on was that
lulesh.chpl on linux32 stumbles across another ssize_t/int64_t
mismatch in IO.chpl.  This is trivial to fix by inserting a cast if
anybody was open to having it go in after the bug fix freeze.  We
fixed a few other instances of this same issue lately and they did not
cause any ripples.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19907 3a8e244f-b0f2-452b-bcba-4c88e055c3ca