Adding a future for the domain address bug.
When a global domain variable is declared using a domain literal, the
broadcast of the address on locale 0 is not done properly.
Specifically, the broadcast call is made before the domain is fully
initialized.

The work around is to explicitly specify the type of the domain.

I have a fix, but given the proximity to the release I didn't feel
confident enough to try and submit it (and then I broke it trying to
get it to be more general).





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19921 3a8e244f-b0f2-452b-bcba-4c88e055c3ca