Restructured the Input and Output chapter of the spec, pursuant the recent group spec review.
This is a minalistic change just to move text around - 
no text was edited - to facilitate future svn diff operations.

The sections:

   \section{Closing Channels and Files}
   \section{Synchronizing Channel and File Data}

were duplicated (once for files, once for channels).

This commit leaves the chapter in an incoherent state, in that
the text does not flow, section vs. subsection distinction
is not proper, some labels are defined twice.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19937 3a8e244f-b0f2-452b-bcba-4c88e055c3ca