Further improvements to the I/O chapter in the spec and README.io.
Includes a good deal of Michael's contributions and feedback.

* Changed the presentation of the implementation-defined I/O types
to address the reference-vs-value-type issue.

The issue is the need to describe, to the reader, certain aspects
of behavior without prescribing the type to be int or record or class etc.

The solution presented here is:
  "A value of the XXX type refers to ..." => semantically a reference type
  "A value of the XXX type defines|represents|contains ... " => a value type

Note: currently, 'file' and 'channel' are semantically reference types
(although they can be implemented with integers, records, etc.),
while 'iohints' and 'syserr' are semantically value types.

In both cases the type is passed by blank intent as an integer, i.e. the
actual value is copied to the formal and cannot be assigned in the function.

Also, iohints, syserr and iostyle are currently 'extern', therefore
I indicated that those types' default values are undefined,
because I remembered seeing a random value in the past and
wanted to be conservative. We can change that as appropriate.

* Added references to README.io under the 'craychapel' environment.

* Removed quite a bit from README.io - those pieces should now
be reflected in the spec proper.

Still left some of the old contents either because it is currently
not in the spec or because it presents certain aspects nicer or differently
than the spec.

We should perhaps sift out more from README.io.

* Moved certain pieces from the spec into README.io:

** Ones in the "implementation-dependent" category.

** When I did not have the time/energy/group participation
to work out/write a spec-quality presentation.

For example, the 'iostyle' type is now presented almost completely
in README.io. Primarily for the above reason, but also because
we haven't settled on what of it we want to require from any Chapel
implementation. Also because we haven't converted some fields to enums
where we have wanted to.

** When Michael provided useful explanations but I did not agree
that this is what the spec should say.

* Moved file/channel synchronization before file/channel close.
Because that's when they logically happen.

* Moved a couple of "write" items before their "read" counterparts.
That's for consistency throughout the chapter.

* Index entries: some new, some \chpl-ified existing ones.

I commented out several \index{<some type>!default value}
because the corresponding text was too small and too close to
\index{<some type>}.

I \chpl-ified the index entries for the 'file' and 'channel' concepts.
That's because it feels tidier to have just one I/O!file entry
than two entries, one \chpl-ified and one not.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19956 3a8e244f-b0f2-452b-bcba-4c88e055c3ca