Reverted r19976: removed 'subdomain' from 'c-other-kwds' in chpl-mode.el for emacs 22.*.
One issue is that r19976 does not achieve its stated purpose.
'subdomain' is not highlighed as a keyword despite being on 'c-other-kwds'
because it is (and has been) also on the 'c-primitive-type-kwds' list.

Another issue, from Brad: "I prefer the current [pre-r19976] behavior in that
it's where we want the language to go and it treats all types similarly."


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19985 3a8e244f-b0f2-452b-bcba-4c88e055c3ca