Ensure BlockCyclicDim compiles for domains with idxType=uint.
It should also compile for idxType=u/int(32).
It wouldn't compile for smaller idxTypes because I couldn't create
ranges with those idxTypes.

Expanded the tmath.chpl test with the newly-added mulP2/divP2/remP2
functions and with testing on each of the four int types.
Disabled prediff sorting of its output (because it is not needed).


To achieve compilation:

* Added more explicit type casts.
(That's "more casts". They already are explicit enough.)

E.g.

-  proc adjustLowIdx()
-    return cycleSizePos - modP2(lowIdxDom, cycleSizePos);

+  proc adjustLowIdx() {
+    const offset = modP2(lowIdxDom, cycleSizePos);
+    return
+      if offset.type == uint then (cycleSizePos:uint) - offset
+      else                        cycleSizePos - offset;
+  }

* Replaced some *, /, % with functions that allow mixing
uint(64) and int(64) arguments. Luckily, when we have such
an occasion, the int(64) argument is expected to be positive,
so I cast it to uint(64) and that preserves any value it may have.

Note: for divP2 and remP2, it is safe to cast the result to
the type of the first argument (in the 'else' cases),
because the result's magnitude will be the same or smaller.
For mulP2, the cast in the 'else' case might not be necessary.

* Simplified the computation of lowIdxAdj. Seems like tending
to the signed-ness of stoIndexT is not necessary.

Also moved _checkFitsWithin to be performed on the result
of the adjustment (which is what we are going to cast),
not on the starting value.

Recall that this check ensures that the value of 'lowIdx', which I
forced to be an 'int' (for simplicity, so that BlockCyclicDim needs
not be generic), is suitable for us to be used with the given domain's
idxType. Having lowIdxAdj as an 'int', or 'uint' if idxType==uint,
instead of running this check, seems to be a valid alternative, too.

The amount of changes that needed to be made to support idxType=uint
(and especially the work it took to arrive at these changes)
is due to binary arithmetic operators (+ - * / %) currently
not supporting the combination of uint(64) and int(64).

Alternative approaches include:

* Do not provide BlockCyclicDim for domains with idxType=uint.
The users need to use idxType=int instead.

* Allow the arithmetic operators to handle the uint+int combination.

* Perform all internal computations/store internal data in 'int(64).
Cast from/to idxType at the interface with the user code.

* If it is important to support uint(64) indices greater than 2^63,
follow the previous bullet but use uint(64) instead of int(64)
when idxType==uint(64).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@19995 3a8e244f-b0f2-452b-bcba-4c88e055c3ca