Fixed a bug in BlockCyclicDim when used for uint domains. Beefed up a test of BlockCyclicDim (t8.chpl).
The bug (not explicitly documented) was that the leader iterator
on DimensionalDist2D with a BlockCyclicDim specifier would not
yield anything in certain cases, especially when the range.low
in the block-cyclic dimension was smaller than adjLowIdx and
idxType was uint. That test case is now part of t8.chpl.

When the param 'zbased' is true, the fix shifts all operations
in BlockCyclic1locdom.dsiMyDensifiedRangeForSingleTask1d()
up by by 'adjLowIdx'. This eliminates the wraparound for the
case where _dsiIndicesOnCycLoc() converts user indices to
zero-based indices by subtracting the 'adjLowIdx' from the former.
Since all that matters is the result of densify(), everything
still works - because
  densify(range, wholeR) == densify(range+adjLowIdx, wholeR+adjLowIdx)

zbased must be 'true' for unsigned idxType. We leave it 'false'
(i.e. all computations proceed like before this commit)
for signed idxType, just to show off the alternative when it is correct.

Also implemented a function to serve as an explicit cast between
range types (e.g. range(int(64)) to range(int(32))), because
such casts are currently not provided.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20001 3a8e244f-b0f2-452b-bcba-4c88e055c3ca