Remove support for pragma "inline"
[reviewed by Tom]

Now that 'inline' has been made into a keyword, it seems redundant to
support it as a pragma as well.  This commit removes the ability to
specify it as a pragma, and rewrites existing code that used the
pragma to use the 'inline' keyword instead.  In brief, my arguments
for getting rid of the pragma are: having two ways to do it is
confusing ("do they imply different inlining semantics?"); the pragma
way is uglier; we haven't exposed pragmas to the user.

On Tom's recommendation, and as a separate change to keep the diffs
cleaner, I'm going to reformat cases that were:

    pragma "inline"
    proc foo(...)

and are now:

    inline
    proc foo(...)

into:

    inline proc foo(...);

which I think is preferable stylistically.  Make some noise if you
don't like this proposal.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20007 3a8e244f-b0f2-452b-bcba-4c88e055c3ca