Added a test of simple aliases.
It tests (a) that multiple and transitive aliasing works, and
(b) that memory is cleaned up nicely.

(a) does not work, so it's a future.

(b) works only for the default distribution, so it is forced in .compopts.

ALSO for multi-locale executions it may crash with sigfault even before
"start" is printed out. This is a known issue with the Memory module, see

  test/memory/vass/crash-memory-module-1.future


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20009 3a8e244f-b0f2-452b-bcba-4c88e055c3ca