Adding CHPL_MAKE_HOME definition
The gen_release script currently fails to tag source and Makefiles
with copyright information because when we refactored the Makefiles'
use of CHPL_MAKE_HOME, we broke this Makefile without noticing it.

This suggests that the gen_release script should be strengthened
to report errors if any of its sub-commands fail (my fault almost
certainly).  To date, we have relied on human inspection to notice
these failures, or for them to ripple downstream into a bigger
failure.  Failing to tag copyrights unfortunately has no such
downtream impact.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20011 3a8e244f-b0f2-452b-bcba-4c88e055c3ca