Another CHPL_MAKE_HOME fix
This is another Makefile that we call explicitly when building the
release which broke because CHPL_MAKE_HOME was not set.  Adding
a set of it.

Arguably, we should do this for all subdirectories of compiler/
(and runtime/) in order to support building a directory in isolation
(for quicker turnaround times), but at this point I'm just trying
to do the minimal changes to get the release working again.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20012 3a8e244f-b0f2-452b-bcba-4c88e055c3ca