SSCA2 kernel#2 rewrite/optimization
[reviewd by Vass and Sung]

Kernel 2 of SSCA#2 effectively computes a reduction, but was using a
sync variable to do so which results in a lot of excess communication
and pounding on a sync variable.  This change switches to using a
reduction which results in significant speedup (8 sec -> 0.3 sec for
the 2-locale performance test on XE for example) and communication
reduction (as indicated by the updated .good file or the comm diagnostics
test).

More could be done to improve this: The subsequent search for all
heaviest edges could be folded in as part of the reduction using a
user-defined reduction (I left it as-is based on the argument that
there wouldn't be many heaviest edges, so probably it wouldn't be that
bad; and out of laziness).  There are also standard throttling issues
(for this code) w.r.t. nested loops that I effectively ignored.  My
main goal was to remove kernel 2 as the bottleneck for scalability,
so hopefully I've done that (though I haven't tested on larger runs
yet, so can't say for sure).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20016 3a8e244f-b0f2-452b-bcba-4c88e055c3ca