Simplified the implementation of the array alias declaration 'var A => expr;' plus accompanying changes: * removed dsiReindex() from DimensionalDist2D * un-futurized simple-alias-mem-use.chpl * updated .good with reduced or changed comm counts
---

Upon 'var A => expr;' where 'expr' is an array, we really want just to take
the pointer to the global array descriptor (a subclass of BaseArr) for 'expr'
and put it in A. Parser converts that statement to a call to _array.newAlias()
which used to invoke dsiReindex() instead. Now newAlias() is this simple:

  record _array {...
    proc newAlias() {
      var x = _value;
      return _newArray(x);
    }
  ...}

The only thing that needs to happen is that array's reference count increment.
Turns out that it still happens, like before, within _newArray(), which does
autocopy upon its return statement. The autocopy increments the ref count.

I will also un-futurize my recent test, simple-alias-mem-use.chpl, which
shows that all memory is deallocated properly when aliasing arrays
mapped using the default distribution.

---

dsiReindex() from DimensionalDist2D is no longer useful. It used to be
there to implement aliasing for DimensionalDist2D-mapped arrays. With
this change, dsiReindex() is no longer invoked for that purpose.
dsiReindex() is still needed to implement reindexing, i.e.
'var A: [...] => expr;', but the one in DimensionalDist2D did not
handle that more general case anyway. So I removed it.

---

This test now passes, so it is no longer a future:
  distributions/robust/arithmetic/reindexing/simple-alias-mem-use.chpl

Warning/reminder: it fails intermittently with comm=gasnet. See
  test/memory/vass/crash-memory-module-1.future

---

Overall comm counts (in those tests that test them) have gone down.
.good have been updated correspondingly.
For studies/ssca2/performance, also updated .prediff. 'cause needed it.

One exception: in each of the following tests, some number of
gets are replaced by an equal number of puts (30 for Alias, 108 for Slice).
The test are in
  test/distributions/robust/arithmetic/performance/
    reduceAlias.chpl
    reduceSlice.chpl
    assignAlias.chpl
    assignSlice.chpl
when run with
  -sdistType=DistType.replicated


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20020 3a8e244f-b0f2-452b-bcba-4c88e055c3ca