Augment SSCA#2 in release/examples to print the elapsed times for Edge Generation and the entire run; include these in perfkeys.
This also adds 3 new graphs, plotting the times for:
* Kernel 2, Kernel 3
* Edge Generation, Kernel 1
* the entire run

We could probaly factor out code to start/stop the timer
and print the timing into a couple of helper functions.

We could also collect those timings in an array or something.
That would let us see how much of the total time we are not accounting for.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20023 3a8e244f-b0f2-452b-bcba-4c88e055c3ca