Create perfdat directory if it does not exist.
If you are running --performance without CHPL_TEST_PERF_DIR set, this
change will create a perfdat directory in $CHPL_HOME/test (if it
doesn't exist) to use for the spec tests performance test.

This worked for the nightly testing because we always set
CHPL_TEST_PERF_DIR to a directory that already exists.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20028 3a8e244f-b0f2-452b-bcba-4c88e055c3ca