Add a communication layer pre-launch hook.
[Reviewed by Sung.]

This adds a communication layer pre-launch hook and changes the launcher
to call that right before it launches a program.  The hook is optional;
if the specific comm layer doesn't define it, then it is nil.

The specific file changes are:

runtime/src/main_launcher.c
  Include the new declarations file chpl-comm-launch.h.  Call the new
  optional communication layer hook CHPL_COMM_PRELAUNCH().

runtime/include/chpl-comm-launch.h
  Provide a default nil definition of CHPL_COMM_PRELAUNCH(), which will
  take effect if the specific communication layer doesn't define it.

runtime/make/Makefile.runtime.head
  Add runtime/include/comm/$(CHPL_MAKE_COMM) to the set of directories
  searched for include files when building the launcher, so that if the
  communication layer provides a chpl-comm-launch.h, we pick that up in
  preference to the regular runtime/include one.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20031 3a8e244f-b0f2-452b-bcba-4c88e055c3ca