test/arrays/diten/arrayPerformance-2d.graph Changed title to "Execution Time", to make it clear that less is better.
    test/types/range/hilde/align64.chpl
Changed type of maxS to int(32) to carry forward the intent of the test.

    modules/dists/ReplicatedDist.chpl
Added a description of the component classes in the module header comment.

    compiler/passes/scopeResolve.cpp
Added some comments about constructor fixup. 

    compiler/resolution/generics.cpp
Adjusted indentation.

    compiler/resolution/functionResolution.cpp
Added !developer tests so that internal names are printed if CHPL_DEVELOPER is
true.

    compiler/optimizations/optimizeOnClauses.cpp
Applied an identity to simplify an apparently complex logical expression:
  a!b v !ab v !a!b <==> !(ab)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20032 3a8e244f-b0f2-452b-bcba-4c88e055c3ca