Pushed 'rmatalt' data structure into studies/ssca2/vass, mostly from r19770. In addition, implemented  removal of duplicate edges.
* Removal of duplicate edges is done by running InsertionSort then
traversing the list shifting edges to the left over duplicates if any.
If boundsChecking is enabled, a VerifySort is done to ensure
correct ordering.
The code is based on the like-named functions in the Sort module,
with an added "payload" array.

TODO: also port quicksort (will need that for larger scales).
Could also modify the sorting code so it eliminates duplicates as it goes.

* Removed John's comment that "The list of neighbors is really just a set"
because it motivates implementing the list via associative domain/array,
which rmatalt does not use.

* While at it, did these renamings, which are internal to RMAT code:
    row_struct -> VertexData (no longer generic)
    Weight     -> edgeWeights

* Added the timing of graph allocation so we can keep an eye on it,
and a corresponding entry to .perfkeys.

* At the moment VertexData includes auxiliary data not used after Kernel 1.
TODO: move them to a separate array that can be deallocated after Kernel 1.

* The total number of duplicate edges is not counted.
For now I just print out 0; could change that.

* The histogram of the number of outgoing edges is now suppressed
when DEBUG_GRAPH_GENERATOR is off. This shortens significantly
the output of performance runs.
Instead, always print the max number of outgoing edges.

* .good also changes the order of edges in each vertex's neighbor list.
The set of edges remains the same. Before they were printed in the
associative-domain (i.e. hashtable) order. Now they are sorted
by the destination vertex ID. (Sorting is the artifact of duplicate removal
and does not seem to affect anything.)

At some point the set of edges will likely change as
edge generation and graph construction become more parallel.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20040 3a8e244f-b0f2-452b-bcba-4c88e055c3ca