First iteration on the improvement of bulk comms
M       modules/internal/DefaultRectangular.chpl

In the previous commit we used doiCanBulkTransfer to check
whether or not both doiBulkTransfer or doiBulkTransferStride were
feasible. Now we have fixed that. For DefaultRectangular,
doiBulkTransfer can be more efficient than doiBulkTransferStride: if
no stride is needed, it may be better to use chpl_comm_get than
chpl_comm_gets. But doiBulkTransfer is much more
restrictive. Therefore we now distinguish between doiCanBulkTransfer
and doiCanBulkTransferStride functions. The former is still more
restrictive than necessary and we have to think again in the
conditions to go the stride way or the non-stride one.

M       modules/internal/ChapelArray.chpl

We also separate chpl__useBulkTransfer and chpl__useBulkTransferStride.
In proc =(a: [], b), we first try to do the doiBulkTransferStride,
but if it is not applicable we then try the doiBulkTransfer. Not sure if it



git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/bulkComms@20042 3a8e244f-b0f2-452b-bcba-4c88e055c3ca