Attempt to remove spurious Mismatched CHPL_HOME warnings
Uses stat to check device and inode numbers
and squelches the warning if these are the same
for a the script. Note that some Windows filesystems
will always return 0, so in some situations
the warning may not be issued even when one
might wish it was; but lately our problem has
been that it is issued too often, so this
should help - even in cases with symbolic
links.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20046 3a8e244f-b0f2-452b-bcba-4c88e055c3ca