Bug fix for 32-bit platforms for strings in --no-local executions
Rafael Asenjo filed a bug pointing out that we core dump on 32-bit
platforms for GASNet executions (or more generally, --no-local
executions).  Tom diagnosed it as being due to an int32_t vs. int64_t
mismatch due to a case that I failed to catch when converting default
ints to 64-bits.  This was going unnoticed on 64-bit platforms (we
believe because of 64-bit field alignment accidentally saving my
bacon).  This commit fixes the issue.  It also makes me wonder what
other similar cases might be lurking undetected...

Other TODOs that we discussed along with this fix:

- commenting the use of the *_DEFAULT type widths in the compiler
  to clarify that they represent Chapel default type widths, not
  C (I'm going to do this in a separate commit).

- using a typedef and corresponding primitive type/alias in the
  compiler for locale IDs in order to be able to switch the type
  for all these fields with a minimal number of changes.

- reordering the fields in wide references and strings to always store
  pointer first, locale ID second, string lenth third.  One of the
  motivations is so that if the locale ID and string length are both
  32 bits again in the future, they will pack better in a 64-bit
  alignment; another is so that references, wide references, and
  strings all share a common prefix (today wide references and strings
  do).

- having the compiler initialize all three fields of a wide string
  rather than just the first two.  valgrind suggests that we aren't
  performing any reads-before-writes due to this currently, but it
  would be cleaner/less weird.

(Then again, it occurred to me this morning that Michael's rewriting
strings, so it may be that the final two bullets will become obsolete
soon).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20051 3a8e244f-b0f2-452b-bcba-4c88e055c3ca