In studies/ssca2/vass, worked around the bug described in r20052, and enabled torus testing again by reversing r20041.
The workaround is to move the entire declaration of 'record VertexData' from
within proc  generate_and_analyze_associative_array_RMAT_graph_representation
to right above it.
Or, the way 'diff' shows it, move the beginning of that proc right after
the 'record VertexData' declaration.

No code changes in this commit other than this move.
Leaving the indentation unchanged to emphasize that it's a workaround.
Alternatively, could make both VertexData and Associative_Graph
available at the top module scope. But that is not motivated at the moment.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20054 3a8e244f-b0f2-452b-bcba-4c88e055c3ca