Adding setchplenv.fish to release
Roald noted this morning that the release tarball does not include
setchplenv.fish, unfortunately.  This was due to a bug in the
gen_release script that hadn't occurred to me which is that currently
we explicitly list all files in the $CHPL_HOME/util/ directory that
should be included in the release (because, currently, our other main
option is to include a directory's full contents, which includes its
subdirectories).

Glancing quickly through the directory, it appears that the
update-sources script also has not been included in releases (or at
least, the most recent one), suggesting that nobody ever tried to use
it from a release.  I didn't add it here, believing that its time of
usefulness has probably passed.

A TODO would be to update the gen_release script to contain a set of
directories whose file contents should be included but subdirectories
should not.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20062 3a8e244f-b0f2-452b-bcba-4c88e055c3ca